NAV-9387 Integrasjon mot Pensjon PESYS * Endrep pathvariable år til int (slipper cast i sql)

+++ src/main/kotlin/no/nav/familie/ba/sak/ekstern/pensjon/PensjonService.kt
@@ -41,7 +41,7 @@ class PensjonService(
         return barnetrygdMedRelaterteSaker.plus(barnetrygdTilPensjon).distinct()
     }
 
-    fun lagTaskForHentingAvIdenterTilPensjon(år: Long): UUID {
+    fun lagTaskForHentingAvIdenterTilPensjon(år: Int): UUID {
         val uuid = UUID.randomUUID()
         taskRepository.save(HentAlleIdenterTilPsysTask.lagTask(år, uuid))
         return uuid
